# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HERA 2.0 Mobile API",
    "version": "2.0.0",
    "description": "API untuk integrasi aplikasi mobile HERA (Hydrology Environmental Real-time Analytics) dengan sistem monitoring kualitas air berbasis IoT. Semua endpoint yang dilindungi memerlukan Bearer Token dari Sanctum.\n\n## Cara Penggunaan\n1. Dapatkan token via `POST /api/mobile/login` atau `POST /api/mobile/register`\n2. Klik tombol **Authorize** di atas dan masukkan token\n3. Coba setiap endpoint langsung dari halaman ini",
    "contact": {
      "name": "Tim HERA 2.0 — Politeknik Negeri Lampung",
      "email": "admin@hera.ac.id"
    }
  },
  "servers": [
    {
      "url": "http://103.217.145.187:8000",
      "description": "HERA 2.0 VPS Production"
    },
    {
      "url": "http://127.0.0.1:8000",
      "description": "Lokal Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Masukkan token dari endpoint /api/mobile/login atau /api/mobile/register"
      }
    },
    "schemas": {
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Budi Santoso" },
          "email": { "type": "string", "example": "petugas@hera.ac.id" },
          "role": { "type": "string", "enum": ["petugas", "direksi"], "example": "petugas" }
        }
      },
      "AuthToken": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "example": "1|abc123xyz..." },
          "token_type": { "type": "string", "example": "Bearer" },
          "expires_in": { "type": "string", "example": "30 days" },
          "user": { "$ref": "#/components/schemas/AuthUser" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Pesan error" }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "The email has already been taken." },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            },
            "example": { "email": ["The email has already been taken."] }
          }
        }
      }
    }
  },
...
  "paths": {
    "/api/mobile/login": {
      "post": {
        "tags": ["Auth Mobile"],
        "summary": "Login mobile",
        "description": "Autentikasi pengguna mobile dan mengembalikan Sanctum Bearer Token yang berlaku 30 hari. Token ini digunakan di semua endpoint yang dilindungi.",
        "operationId": "mobileLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["email", "password"],
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "petugas@hera.ac.id" },
                  "password": { "type": "string", "format": "password", "example": "password123" },
                  "device_name": { "type": "string", "example": "Samsung Galaxy A54", "description": "Nama perangkat (opsional, untuk identifikasi token)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Login berhasil." },
                    "data": { "$ref": "#/components/schemas/AuthToken" }
                  }
                }
              }
            }
          },
          "401": { "description": "Email atau password salah", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/mobile/register": {
      "post": {
        "tags": ["Auth Mobile"],
        "summary": "Register akun baru",
        "description": "Membuat akun pengguna baru dengan role `petugas` (default). Langsung mengembalikan Bearer Token sehingga pengguna tidak perlu login ulang setelah mendaftar.",
        "operationId": "mobileRegister",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["name", "email", "password"],
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Andi Prasetyo" },
                  "email": { "type": "string", "format": "email", "example": "andi@hera.ac.id" },
                  "password": { "type": "string", "format": "password", "example": "password123", "description": "Minimal 6 karakter" },
                  "device_name": { "type": "string", "example": "Xiaomi Redmi Note 12", "description": "Nama perangkat (opsional)" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registrasi berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Registrasi berhasil. Selamat datang di HERA!" },
                    "data": { "$ref": "#/components/schemas/AuthToken" }
                  }
                }
              }
            }
          },
          "422": { "description": "Validasi gagal (email sudah terdaftar, password < 6 karakter, dsb.)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationError" } } } }
        }
      }
    },
    "/api/mobile/logout": {
      "post": {
        "tags": ["Auth Mobile"],
        "summary": "Logout mobile",
        "description": "Mencabut (revoke) Bearer Token yang sedang aktif. Token tidak dapat digunakan lagi setelah ini.",
        "operationId": "mobileLogout",
        "security": [{ "sanctum": [] }],
        "responses": {
          "200": { "description": "Logout berhasil", "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "boolean", "example": true }, "message": { "type": "string", "example": "Logout berhasil. Token telah dicabut." } } } } } },
          "401": { "description": "Token tidak valid atau sudah kedaluwarsa" }
        }
      }
    },
    "/api/mobile/profile": {
      "get": {
        "tags": ["Auth Mobile"],
        "summary": "Profil pengguna",
        "description": "Mengembalikan data profil pengguna yang sedang login berdasarkan Bearer Token.",
        "operationId": "mobileProfile",
        "security": [{ "sanctum": [] }],
        "responses": {
          "200": {
            "description": "Data profil berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "OK" },
                    "data": { "$ref": "#/components/schemas/AuthUser" }
                  }
                }
              }
            }
          },
          "401": { "description": "Token tidak valid" }
        }
      }
    },
...
  }
}